Move add-data & add-cache from blockdev -> pool
It seems more intuitive to add devices to the pool than to add them
to a blockdev.
