+++ src/stratis_cli/_actions/_physical.py
@@ -19,9 +19,6 @@ from __future__ import print_function
 
 from justbytes import Range
 
-from .._errors import StratisCliEngineError
-from .._stratisd_constants import StratisdErrors
-
 from ._connection import get_object
 from ._constants import SECTOR_SIZE
 from ._constants import TOP_OBJECT
@@ -29,9 +26,6 @@ from ._constants import UNKNOWN_VALUE_MARKER
 from ._data import devs
 from ._data import MODev
 from ._data import ObjectManager
-from ._data import Pool
-from ._data import pools
-from ._data import unique
 from ._formatting import print_table
 from ._util import get_objects
 
@@ -62,6 +56,7 @@ class PhysicalActions():
     """
     Actions on the physical aspects of a pool.
     """
+    # pylint: disable=too-few-public-methods
 
     @staticmethod
     def list_pool(namespace):
@@ -86,43 +81,3 @@ class PhysicalActions():
             ["Pool Name", "Device Node", "Physical Size", "State", "Tier"],
             sorted(tables, key=lambda entry: (entry[0], entry[1])),
             ['<', '<', '>', '>', '>'])
-
-    @staticmethod
-    def add_data_device(namespace):
-        """
-        Add a device to a pool.
-        """
-        proxy = get_object(TOP_OBJECT)
-        managed_objects = ObjectManager.Methods.GetManagedObjects(proxy, {})
-        (pool_object_path, _) = unique(
-            pools(props={
-                'Name': namespace.pool_name
-            }).search(managed_objects))
-
-        (_, rc, message) = Pool.Methods.AddDataDevs(
-            get_object(pool_object_path), {
-                'force': namespace.force,
-                'devices': namespace.device
-            })
-        if rc != StratisdErrors.OK:
-            raise StratisCliEngineError(rc, message)
-
-    @staticmethod
-    def add_cache_device(namespace):
-        """
-        Add a device to a pool.
-        """
-        proxy = get_object(TOP_OBJECT)
-        managed_objects = ObjectManager.Methods.GetManagedObjects(proxy, {})
-        (pool_object_path, _) = unique(
-            pools(props={
-                'Name': namespace.pool_name
-            }).search(managed_objects))
-
-        (_, rc, message) = Pool.Methods.AddCacheDevs(
-            get_object(pool_object_path), {
-                'force': namespace.force,
-                'devices': namespace.device
-            })
-        if rc != StratisdErrors.OK:
-            raise StratisCliEngineError(rc, message)
