+++ src/stratis_cli/_parser/_physical.py
@@ -18,49 +18,6 @@ Definition of block device actions to display in the CLI.
 from .._actions import PhysicalActions
 
 PHYSICAL_SUBCMDS = [
-    ('add-data',
-     dict(
-         help=
-         "Add one or more blockdevs to an existing pool for use as data storage",
-         args=[
-             ('pool_name', dict(
-                 action='store',
-                 help='Pool name',
-             )),
-             ('device',
-              dict(
-                  help='Block devices to add to the pool',
-                  metavar='blockdev',
-                  nargs='+')),
-             ('--force',
-              dict(
-                  action='store_true',
-                  default=False,
-                  help=
-                  "Use devices even if they appear to contain existing data")),
-         ],
-         func=PhysicalActions.add_data_device)),
-    ('add-cache',
-     dict(
-         help="Add one or more blockdevs to an existing pool for use as cache",
-         args=[
-             ('pool_name', dict(
-                 action='store',
-                 help='Pool name',
-             )),
-             ('device',
-              dict(
-                  help='Block devices to add to the pool as cache',
-                  metavar='blockdev',
-                  nargs='+')),
-             ('--force',
-              dict(
-                  action='store_true',
-                  default=False,
-                  help=
-                  "Use devices even if they appear to contain existing data")),
-         ],
-         func=PhysicalActions.add_cache_device)),
     ('list',
      dict(
          help="List information about blockdevs in the pool",
