+++ src/stratis_cli/_parser/_pool.py
@@ -79,4 +79,47 @@ POOL_SUBCMDS = [
          ],
          func=TopActions.rename_pool,
      )),
+    ('add-data',
+     dict(
+         help=
+         "Add one or more blockdevs to an existing pool for use as data storage",
+         args=[
+             ('pool_name', dict(
+                 action='store',
+                 help='Pool name',
+             )),
+             ('device',
+              dict(
+                  help='Block devices to add to the pool',
+                  metavar='blockdev',
+                  nargs='+')),
+             ('--force',
+              dict(
+                  action='store_true',
+                  default=False,
+                  help=
+                  "Use devices even if they appear to contain existing data")),
+         ],
+         func=TopActions.add_data_device)),
+    ('add-cache',
+     dict(
+         help="Add one or more blockdevs to an existing pool for use as cache",
+         args=[
+             ('pool_name', dict(
+                 action='store',
+                 help='Pool name',
+             )),
+             ('device',
+              dict(
+                  help='Block devices to add to the pool as cache',
+                  metavar='blockdev',
+                  nargs='+')),
+             ('--force',
+              dict(
+                  action='store_true',
+                  default=False,
+                  help=
+                  "Use devices even if they appear to contain existing data")),
+         ],
+         func=TopActions.add_cache_device)),
 ]
